disable tilde subscripting to fix manpage
Asciidoc uses tildes to subscript and carets to superscript text.
The first tilde can be escaped to fix this, but looks weird because
the second tilde may not be escaped.

It doesn't seem like subscripting will be used in the man page so it
makes sense to disable it altogether.