Add swaybg executable
This will let users set their background to something other than that
cringy demoscene thing wlc has now. It's also going to be the first
wayland client written for sway, so I picked an easy thing to work on.
We'll have to figure out how to indicate that it's a special view.